correction des differentes fonctions pour prendre en compte les nombres

+++ fonction/operation.py
@@ -1,21 +1,21 @@
 def addition(a,b):
-    a = int(a)
-    b = int(b)
+    a = float(a)
+    b = float(b)
     return a+b
 
 def soustraction(a,b):
-    a = int(a)
-    b = int(b)
+    a = float(a)
+    b = float(b)
     return a-b
 
 def multiplication(a,b):
-    a = int(a)
-    b = int(b)
+    a = float(a)
+    b = float(b)
     return a*b
 
 def division(a,b):
-    a = int(a)
-    b = int(b)
+    a = float(a)
+    b = float(b)
     if b != 0:
         return a/(b)
     else:
